Some improvements, added some format variations
