+++ NuXtractor/Scenes/Scene.cs
@@ -1,8 +1,5 @@
-﻿using SixLabors.ImageSharp;
-using System.Collections.Generic;
-using System.IO;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 
 namespace NuXtractor.Scenes
 {
@@ -30,46 +27,6 @@ namespace NuXtractor.Scenes
         public Scene()
         {
             Objects = new List<SceneObject>();
-        }
-
-        public async Task ArchiveAsync(string dir)
-        {
-            Directory.CreateDirectory(dir);
-
-            var mtlPath = Path.Combine(dir, "materials.mtl");
-
-            using (var writer = File.CreateText(mtlPath))
-            {
-                writer.AutoFlush = true;
-                foreach (var material in Materials)
-                {
-                    await material.WriteToMTLAsync(writer);
-                }
-            }
-
-            var texDir = Path.Combine(dir, "textures");
-            Directory.CreateDirectory(texDir);
-
-            foreach (var texture in Textures)
-            {
-                var image = await texture.ReadImageAsync();
-
-                var texPath = Path.Combine(texDir, $"texture_{texture.Id}.png");
-                await image.SaveAsPngAsync(texPath);
-            }
-
-            var objPath = Path.Combine(dir, "scene.obj");
-            using (var writer = File.CreateText(objPath))
-            {
-                writer.AutoFlush = true;
-
-                await writer.WriteLineAsync("mtllib .\\materials.mtl");
-
-                foreach (var obj in Objects)
-                {
-                    await obj.WriteToOBJAsync(writer);
-                }
-            }
         }
     }
 }
